pyproject.toml
[tool.poetry]
name = "goit-web-hw-10"
version = "0.1.0"
description = ""
authors = ["R.Anatolii <tv.weplay@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
mongoengine = "^0.29.1"
pymongo = "^4.10.1"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
scrapy = "^2.11.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
